Added MullerLyer Illusion
Added calculation of error in mm.
Major db Update of MullerLyer Illusion table

This commit adds the Circle classes to the geometry_utils module. The Circle class represents a circle with a center and a radius, and it includes methods for rotation and drawing on a canvas.

The changes also include updating the PoggendorffIllusion class in the gui module to use the Line and Circle classes for drawing lines and circles on the canvas.